Resolve gcc compiler warnings in ARM environment

+++ source/SoundStretch/WavFile.cpp
@@ -450,7 +450,7 @@ int WavInFile::read(float *buffer, int maxElems)
 int WavInFile::eof() const
 {
     // return true if all data has been read or file eof has reached
-    return (dataRead == header.data.data_len || feof(fptr));
+    return ((uint)dataRead == header.data.data_len || feof(fptr));
 }
